src/requirements.txt
fastapi==0.104.1
mangum==0.17.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
pydantic[email]==2.5.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0
uvicorn==0.24.0